Split out ScrollingHitObjectContainer into new file
